==> Part 1 Web Scraping & Data Storage/4. Sectional Title Property Scrape 200.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait, Select
from selenium.webdriver.support import expected_conditions as EC
import pandas as pd

driver = webdriver.Chrome()
driver.get('https://valuation2022.durban.gov.za/') 

driver.switch_to.frame("mainFrame")

valuation_roll_dropdown = WebDriverWait(driver, 10).until(
    EC.presence_of_element_located((By.ID, "drpSearchType"))
)
select = Select(valuation_roll_dropdown)
select.select_by_visible_text("Sectional Title Property")

go_button = WebDriverWait(driver, 10).until(
    EC.element_to_be_clickable((By.ID, "btnGo"))
)
driver.execute_script("arguments[0].click();", go_button)

driver.switch_to.frame("frmSearchCriter")

dropdown_element = WebDriverWait(driver, 15).until(
    EC.presence_of_element_located((By.XPATH, "//table//select[@id='drpVolumeNo']"))
)

dropdown = Select(dropdown_element)
dropdown.select_by_visible_text("01")

search_button = WebDriverWait(driver, 10).until(
    EC.element_to_be_clickable((By.XPATH, "/html/body/form/div[3]/table/tbody/tr[1]/td[2]/table/tbody/tr/td[1]/input"))
)
driver.execute_script("arguments[0].click();", search_button)

driver.switch_to.default_content()


WebDriverWait(driver, 15).until(EC.frame_to_be_available_and_switch_to_it((By.NAME, "mainFrame")))
print("Switched to mainFrame")

  



WebDriverWait(driver, 30).until(EC.frame_to_be_available_and_switch_to_it((By.NAME, "frameSearch")))
print("Switched to frameSearch")

  
table = WebDriverWait(driver, 15).until(
    EC.presence_of_element_located((By.XPATH, "//table[contains(@class, 'Grid') or contains(@border, '1')]"))
    )
print("Table located")


rows = table.find_elements(By.TAG_NAME, "tr")
data = []
for row in rows[2:]:
    cols = row.find_elements(By.TAG_NAME, "td") or row.find_elements(By.TAG_NAME, "th")
    data.append([col.text.strip() for col in cols])

data = [row for row in data if any(cell.strip() for cell in row)]

num_columns = len(data[0])
valid_data = [row for row in data if len(row) == num_columns]
first_200_data = valid_data[:201]    
df = pd.DataFrame(first_200_data[1:], columns=first_200_data[0])

print(df.head(200))
df.to_csv(r'C:\Users\Kams\Desktop\tech-challenge-main\Sectional Title Property Data.csv', index=False)
print("First 200 records saved to 'Sectional_Title_Property_Data.csv'")
