==> Part 1 Web Scraping & Data Storage/6. Data Cleaning And Output.ipynb <==
import pandas as pd #import pandas

dfFull = pd.read_csv(r'C:\Users\Kams\Desktop\tech-challenge-main\Part 1 Web Scraping & Data Storage\Full Title Property Data.csv', delimiter = ',')  #two dataframes from two csvs using the delimiter ','
dfSect = pd.read_csv(r'C:\Users\Kams\Desktop\tech-challenge-main\Part 1 Web Scraping & Data Storage\Sectional Title Property Data.csv', delimiter = ',')

pd.set_option('display.width', 200) #set the display so all items are viewable
pd.set_option('display.max_columns', None)

dfFull.head() #shows the 1st few records in DataframeFull

dfSect.head()#shows the 1st few records in DataframeSect

dfFull.info() #Checking if datatypes are Correct
dfSect.info() 

dfFull.isnull().sum()
dfSect.isnull().sum() #Checking For Null Values

dfFull.fillna(0) #Filling in Nulls with 0
dfSect.fillna(0)

dfFull = dfFull.drop_duplicates() #Deleting Duplicates
dfSect = dfSect.drop_duplicates()

dfFull['First Owner'] = dfFull['First Owner'].replace('R S A', 'RSA') #Fixing Some strings

dfFull #display results
dfSect

dfFull.to_excel(r'C:\Users\Kams\Desktop\tech-challenge-main\Part 1 Web Scraping & Data Storage\Full Title Property Data Cleaned.xlsx', index=False) #Save in xlsx Files
dfSect.to_excel(r'C:\Users\Kams\Desktop\tech-challenge-main\Part 1 Web Scraping & Data Storage\Sectional Title Property Data Cleaned.xlsx', index=False)
