==> Part 1 Web Scraping & Data Storage/9. Store Data In A PostgreSQL Database.ipynb <==
import pandas as pd #Using pandas Dataframes
from sqlalchemy import create_engine #And SQLAlchemy To Create Tables

dfFull = pd.read_excel(r'C:\Users\Kams\Desktop\tech-challenge-main\Part 1 Web Scraping & Data Storage\7. Full Title Property Data Cleaned.xlsx') #dataframe stores Data From xlsx 
dfSect = pd.read_excel(r'C:\Users\Kams\Desktop\tech-challenge-main\Part 1 Web Scraping & Data Storage\8. Sectional Title Property Data Cleaned.xlsx')

engine = create_engine('postgresql://postgres:1234567890@localhost:5432/Valuation 2022 Durban') #Connects to the Database Using the Password(NOT HIDDEN)

dfFull.to_sql('FullTitle Property Data', engine, if_exists='replace', index=False) #Creates Tables If table already Exists Replace It
dfSect.to_sql('SectionalTitle Property Data', engine, if_exists='replace', index=False )
